% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawFiles.R
\name{checkRawFiles}
\alias{checkRawFiles}
\title{check raw files}
\usage{
checkRawFiles(meta, size = TRUE)
}
\arguments{
\item{meta}{metafile downloaded from PRIDE using getProjectMeta function}
}
\description{
extract raw files from meta files of a PRIDE project
}
\examples{
uri <- 'ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2017/01/PXD003701/'
pmeta <- getProjectMeta(uri)
checkRawFiles(pmeta)
}
