% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectMeta.R
\name{getProjectMeta}
\alias{getProjectMeta}
\title{Get meta information about a PRIDE project}
\usage{
getProjectMeta(uri)
}
\arguments{
\item{uri}{uri of PRIDE archive}
}
\description{
Download meta file information based on PRIDE uri
}
\examples{
uri <- 'ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2017/01/PXD003701/'
pmeta <- getProjectMeta(uri)
}
