% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectsPerYear.R
\name{getProjectsPerYear}
\alias{getProjectsPerYear}
\title{Get yearwise projects available at PRIDE database}
\usage{
getProjectsPerYear(year = "2017")
}
\arguments{
\item{year}{numeric/character}
}
\description{
Get PRIDE projects information from PRIDE database based on year
}
\examples{
getProjectsPerYear(year = 2017)
}
