% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawFiles.R
\name{getRawFiles}
\alias{getRawFiles}
\title{Get raw files}
\usage{
getRawFiles(meta, path = NULL)
}
\arguments{
\item{meta}{metafile downloaded from PRIDE using getProjectMeta function}
}
\description{
download raw files
}
\examples{
uri <- 'ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2017/01/PXD003701/'
pmeta <- getProjectMeta(uri)
# caution before downloading please check file size using checkRawFiles function
checkRawFiles(pmeta)
getRawFiles(meta = pmeta)
}
